quiz-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.6.0
pydantic==2.9.2
httpx==0.27.2

grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1